+++ quelware-client/src/quelware_client/testing/instrument_agent_mock/__init__.py
@@ -60,5 +60,13 @@ class InstrumentAgentMock(InstrumentAgent):
     ) -> ResultContainer:
         return ResultContainer()
 
+    async def wait_for_result(
+        self,
+        token: SessionToken,
+        resource_id: ResourceId,
+        timeout_sec: float | None,
+    ) -> ResultContainer:
+        return ResultContainer()
+
 
 __all__ = ["InstrumentAgentMock"]